refactor(gitignore): move package to core/utils/git for reuse
Relocate the gitignore matcher (package gitutils) from the App module
(internal/utils/git) into the core library module (core/utils/git) so
other projects can import it. Package name and API are unchanged; only
the import path moves to github.com/codetrek/haystack/core/utils/git.

- git mv all sources + tests (history preserved)
- add github.com/sabhiram/go-gitignore to core/go.mod (the only new dep;
  core/go.sum updated, core builds standalone under GOWORK=off)
- repoint the two real consumers (indexer/scanner.go, fs/list_files_test.go)
- App keeps go-gitignore directly (still used by internal/utils/simple_filter.go),
  so its go.mod and vendor stay consistent — no App dependency change

Verified: core builds+tests standalone (GOWORK=off) and via go.work -race;
App indexer + fs tests green; workspace go build ./... green.
