+++ core/utils/git/ignore.go
