+++ core/utils/git/ignore_test.go
