+++ core/utils/git/pattern_match.go
