+++ core/utils/git/pattern_match_test.go
