+++ internal/server/indexer/scanner.go
@@ -7,11 +7,11 @@ import (
 	"sync"
 	"time"
 
+	gitutils "github.com/codetrek/haystack/core/utils/git"
 	"github.com/codetrek/haystack/internal/core/workspace"
 	"github.com/codetrek/haystack/internal/shared/running"
 	"github.com/codetrek/haystack/internal/utils"
 	fsutils "github.com/codetrek/haystack/internal/utils/fs"
-	gitutils "github.com/codetrek/haystack/internal/utils/git"
 )
 
 type GitIgnoreFilter struct {
